move setcriterion to loss
