+++ easycv/models/loss/__init__.py
@@ -4,3 +4,4 @@ from .focal_loss import FocalLoss, VarifocalLoss
 from .iou_loss import GIoULoss, IoULoss, YOLOX_IOULoss
 from .mse_loss import JointsMSELoss
 from .pytorch_metric_learning import *
+from .set_criterion import *
